Move commands to their own package
