+++ server/src/main/kotlin/net/stellarica/server/StellaricaServer.kt
@@ -5,7 +5,7 @@ import mu.KotlinLogging
 import net.minecraft.resources.ResourceLocation
 import net.stellarica.server.craft.starship.InterfaceListener
 import net.stellarica.server.craft.starship.Starship
-import net.stellarica.server.craft.starship.StarshipCommands
+import net.stellarica.server.command.StarshipCommands
 import net.stellarica.server.material.custom.CustomMaterialCommands
 import net.stellarica.server.material.custom.block.CustomBlockHandler
 import net.stellarica.server.material.custom.block.CustomBlocks
@@ -17,7 +17,7 @@ import net.stellarica.server.multiblock.MultiblockHandler
 import net.stellarica.server.multiblock.Multiblocks
 import net.stellarica.server.networking.BukkitNetworkHandler
 import net.stellarica.server.networking.ModdedPlayerHandler
-import net.stellarica.server.util.DebugCommands
+import net.stellarica.server.command.DebugCommands
 import org.bukkit.Bukkit.getPluginManager
 import org.bukkit.NamespacedKey
 import org.bukkit.entity.Player
